Server : basic LogicCell serialization
