+++ Assets/Prefabs/MustHaves/ClientNetworking.prefab
@@ -43,7 +43,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: ad6e58ddc91bc2e41a1691bec489cc39, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  m_udpClient: {fileID: 1584834601242082865}
+  m_UDPClient: {fileID: 1584834601242082865}
+  m_TCPClient: {fileID: 5214724981417358588}
   m_playerSettings: {fileID: 11400000, guid: b8d133b304cd13c40b0e4e4a37bb5595, type: 2}
   m_physicsScene: proto_move_client
   ConnectButtonEvent:
