+++ Assets/Scripts/Common/Serialization.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3681aa149ae6dc946a9ff732b58c3cf1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
