+++ Assets/Scripts/Common/Serialization/Serializable.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 62af5f924ecd511429a3a6521f7e84c9
+guid: 2c0730452b0fad1439f3711071ae83c4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
