+++ Assets/Scripts/Common/Serialization/Serialization.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d0aef6a8a84671946b14027cfc6808a5
+guid: 7ef70e033f50bb64dad1da4423d93b09
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
