+++ Assets/Scripts/Common/World/CellType/WallCell.cs
@@ -12,5 +12,10 @@ namespace ubv.common.world.cellType
         {
             IsWalkable = false;
         }
+
+        protected override byte SerializationID()
+        {
+            return (byte)serialization.ID.BYTE_TYPE.LOGIC_CELL_WALL;
+        }
     }
 }
